# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/chats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Список моих чатов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/group": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Создать групповой чат",
                "parameters": [
                    {
                        "description": "название и участники",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_modules_chat_domain.CreateGroupChatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/private": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Создать приватный чат",
                "parameters": [
                    {
                        "description": "id собеседника",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_modules_chat_domain.CreatePrivateChatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Поиск чатов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "строка поиска",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Информация о чате",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Обновить чат",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "поля title, avatar_url",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}/leave": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Покинуть чат",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/chats/{chat_id}/mute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Заглушить уведомления чата",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Включить уведомления чата",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}/participants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Участники чата",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Добавить участника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "поле user_id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}/participants/{user_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Удалить участника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id участника",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}/participants/{user_id}/demote": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Снять админа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id участника",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/chats/{chat_id}/participants/{user_id}/promote": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Назначить админа",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id чата",
                        "name": "chat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id участника",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jeogram_messenger_internal_modules_chat_domain.CreateGroupChatRequest": {
            "type": "object",
            "required": [
                "participant_ids",
                "title"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "maxLength": 512
                },
                "participant_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 128
                }
            }
        },
        "github_com_jeogram_messenger_internal_modules_chat_domain.CreatePrivateChatRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_jeogram_messenger_internal_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_jeogram_messenger_internal_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_jeogram_messenger_internal_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}